+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 1.3.6 (unreleased)
 ------------------
 
+- #2321 Improve the creation process of AT content types (#2150 port)
 - #2320 Performance: prioritize raw getter for AllowedMethods field (#2149 port)
 - #2308 Rely on fields when validating values on sample creation (#2307 port)
 - #2249 Fix APIError: None is not supported (upgrade 1.3.6)